Move all Git-related operations to GitVersion
GradleUtils 2.3 was 90% JGit and Git-related operations, which are all
now housed in the framework-agnostic GitVersion. Projects that do not
opt-in to using GitVersion or having subprojects are not affected.

There are a small amount of binary breaking changes in this update, but
it should not affect Gradle buildscripts (the primary consumer of this)
as they are dynamically compiled and can use Groovy's funky syntax to
get what they need.