Allow lenient GitVersion creation (until 3.0)
Subprojects that do not need to use their own version will use the root
project's version. For now this is handled by catching failure of the
Git Version creation on the subproject, and trying to create one using
the root project instead.

Management of Git Version will be refined in GradleUtils 3.0. Consider
the usage of it an incubating feature until then.