Fixed issues with Gradle property decoration
Any public or protected method that has the getter naming convention
(getProperty) is decorated by Gradle when the class is instantiated. A
lot of things in the shared package are largely internal and do not need
modifications from Gradle, so they have been renamed without the "get"
keyword to avoid that. For non-tasks, the `@Internal` annotation has no
effect.