Split off shared base into its own subproject
Note: This is NOT another plugin.

This week, a lot of work has been spent merging the common code between
FG7, FD7, and GU3 into one shared package that all of them can extend
off of. Originally this was just a package within GU3 itself, but this
will inevitably cause problems with plugins that may need to interface
with GU3 itself (not the shared base).

My solution was to create 'gradleutils-shared', a project that holds
only this common code and nothing else. The point is for our Gradle
projects to shadow in this code so that it doesn't have an arbitrary
requirement on GU3 being present in the classpath. I am publishing
'gradleutils-shared' in a way so that the shadow jar is the only one
that can be consumed. Project dependencies cannot read module files,
however, so the GU3 buildscript itself has a workaround for it.