Lower targets for gradleutils-shared
Need to target Gradle 8.13 because ForgeDev is currently on 8.14.1 right
now, and FG6 will not run on Gradle 9. I need FG6 in the workspace with
Forge for now since it acts as a reference point to help me move the
patcher plugin application. Once ForgeDev is free of FG6, it will be
updated to Gradle 9.

Also, got a tip from the Gradle Slack that worker daemons
(`WorkerExecutor`) can still run in Java 8 on Gradle 9, so making this
shared code target Java 8 will be better long-term.