Stabilize publication promotion
Publication promotion (also known as artifact promotion) promotes a
published publication on the Forge files site to become the "latest"
version it displays. It also updates the files site and makes it aware
of the new version of the project that was just pushed to the maven.

This is designed for internal use only by Forge projects, or for
projects that have trusted access to the Forge maven. However, it is not
a requirement for all Forge projects to be promoted on the files site.

As a reminder, promotion is requested like this in your buildscript
(Groovy DSL):

```groovy
publishing {
    repositories {
        maven gradleutils.publishingForgeMaven
    }

    publications.register('pluginMaven', MavenPublication) {
        gradleutils.promote(it)

        // ...
    }
}
```