Add flag to set publication artifact ID from base.archivesName
Apologies for forgetting how unbelievably inconsistent Gradle can be.
The project's `base.archivesName` is used as the base name for archives,
but that name isn't used for publications even though it's used for the
archives. It's deep in internals, but publication name is set by
`project.name`. When writing our buildscripts, I was under the
impression that `base.archivesName` was used for the artifact ID for
publications. Thankfully, most of our projects already use the project
name as the artifact ID, so there has been no impact.

I've added it in GradleUtils but it's disabled by default. You can use
it by setting this flag in your `gradle.properties`:
```properties
net.minecraftforge.gradleutils.publishing.use-base-archives-name=true
```