Allow usage of #launcherFor without the 'java' plugin
Not all Gradle plugins are required to apply the 'java' plugin. This
means that for plugins which need a Java launcher for a task and/or
JavaExec(Spec), we need to be flexible to avoid locking in projects into
needing that plugin.

I've locked down the SharedUtil methods for getting Java launchers.
#launcherFor requires Project and #launcherForStrict requires
JavaToolchainService (provided by JavaExecSpec, or can be injected at
project configuration). If no Java toolchain or 'java' plugin,
`JavaLanguageVersion#current` will be used to compare to the required
Java version to run.

As a reminder, breaking changes are permitted for `gradleutils-shared`
as it must only ever be shadowed by plugins that use it.