Remove Gradle version attribute from the shared library
The Gradle version attribute is primarily used to enforce that a
specific Gradle version is used for a plugin. Since the GradleUtils
Shared library is not a plugin, this causes issues if added as a
dependency because Gradle will indescriminately check the current
daemon's running Gradle version even if it has nothing to do with its
application.

Since this is an internal library, I am not going to enforce the Gradle
API dependency constraint on this library, especially since the Gradle
API artifact is not centralized or declares any capabilities. But, if
non-Forge consumers wish to use GradleUtils Shared for their plugins, it
is up to them to make sure their implementation Gradle version is high
enough. As if right now, it is 8.13.