Honor system property to suppress Gradle API
This applies a workaround to gradle/plugin-portal-requests#260 when the
Shadow plugin is applied where the local Gradle API is added to the
`compileOnly` configuration even if the
`org.gradle.unsafe.suppress-gradle-api` system property is used. Which,
we do.

This workaround is written in a way that even after this is fixed by the
Plugin Publish plugin, it won't break anything. It just helps remove
some clutter from the buildscript.