Remove overloading GradleUtilsExtension#unpack methods
These methods cause a GroovyRuntimeException as the runtime invoker
cannot choose between some of the overloading methods. This bug is rare
and only happens in composite builds, but that unfortunately affects Git
Version.

This is unfortunately a binary breaking change but does not affect
buildscripts. Plugins that shadow GradleUtils Shared are unaffected, and
there shouldn't be any plugins that are compiled to use GradleUtils at
runtime.